trackpoint: Fixup the drift correction after lego.
Drift correction got broken with lego.  This fixes it.
